+++ Example/Non-Card Payment Examples/Non-Card Payment Examples/CryptoExampleViewController.swift
@@ -1,0 +1,116 @@
+//
+//  CryptoExampleViewController.swift
+//  Non-Card Payment Examples
+//
+//  Created by Eric Zhang on 11/20/24.
+//
+
+import Stripe
+import UIKit
+
+class CryptoExampleViewController: UIViewController {
+    @objc weak var delegate: ExampleViewControllerDelegate?
+    var inProgress: Bool = false {
+        didSet {
+            navigationController?.navigationBar.isUserInteractionEnabled = !inProgress
+            payButton.isEnabled = !inProgress
+            inProgress
+                ? activityIndicatorView.startAnimating() : activityIndicatorView.stopAnimating()
+        }
+    }
+
+    // UI
+    lazy var activityIndicatorView = {
+        return UIActivityIndicatorView(style: .gray)
+    }()
+    lazy var payButton: UIButton = {
+        let button = UIButton(type: .roundedRect)
+        button.setTitle("Pay with Crypto", for: .normal)
+        button.addTarget(self, action: #selector(didTapPayButton), for: .touchUpInside)
+        return button
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .white
+        title = "Crypto"
+        [payButton, activityIndicatorView].forEach { subview in
+            view.addSubview(subview)
+            subview.translatesAutoresizingMaskIntoConstraints = false
+        }
+
+        let constraints = [
+            payButton.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            payButton.centerYAnchor.constraint(equalTo: view.centerYAnchor),
+
+            activityIndicatorView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            activityIndicatorView.centerYAnchor.constraint(equalTo: view.centerYAnchor),
+        ]
+        NSLayoutConstraint.activate(constraints)
+    }
+
+    @objc func didTapPayButton() {
+        guard STPAPIClient.shared.publishableKey != nil else {
+            delegate?.exampleViewController(
+                self,
+                didFinishWithMessage: "Please set a Stripe Publishable Key in Constants.m"
+            )
+            return
+        }
+        inProgress = true
+        pay()
+    }
+}
+
+// MARK: - Crypto
+extension CryptoExampleViewController {
+    @objc func pay() {
+        // 1. Create an Crypto PaymentIntent
+        MyAPIClient.shared().createPaymentIntent(
+            completion: { (_, clientSecret, error) in
+                guard let clientSecret = clientSecret else {
+                    self.delegate?.exampleViewController(self, didFinishWithError: error)
+                    return
+                }
+
+                // 2. Confirm the payment and redirect the user to Crypto
+                let paymentIntentParams = STPPaymentIntentParams(clientSecret: clientSecret)
+                paymentIntentParams.paymentMethodParams = STPPaymentMethodParams(
+                    crypto: STPPaymentMethodCryptoParams(),
+                    billingDetails: nil,
+                    metadata: nil
+                )
+                paymentIntentParams.returnURL = "payments-example://safepay/"
+
+                STPPaymentHandler.shared().confirmPayment(
+                    paymentIntentParams,
+                    with: self
+                ) { (status, _, error) in
+                    switch status {
+                    case .canceled:
+                        self.delegate?.exampleViewController(
+                            self,
+                            didFinishWithMessage: "Cancelled"
+                        )
+                    case .failed:
+                        self.delegate?.exampleViewController(self, didFinishWithError: error)
+                    case .succeeded:
+                        self.delegate?.exampleViewController(
+                            self,
+                            didFinishWithMessage: "Payment successfully created."
+                        )
+                    @unknown default:
+                        fatalError()
+                    }
+                }
+            },
+            additionalParameters: "supported_payment_methods=crypto"
+        )
+    }
+}
+
+extension CryptoExampleViewController: STPAuthenticationContext {
+    func authenticationPresentingViewController() -> UIViewController {
+        self
+    }
+}